Restructured project.
Updated the deployment to use fabric and invoke.

Created a set of invoke tasks that are organised by the service or
feature it installs/manages. That will make it much easier to add
new features.

Templates are rendered in two steps. First the cookiecurrent values
are filled in. Then during the deployment the values from fabric.json
are used to fill out template variables (escaped in the first step).
That makes it much easier to deploy projects with different layouts.

Added features from various sources.