No need to restart celery when deploying for the first time.
Remove the task to restart celery and celery beat. When supervisor
loads the configuration using the restart command all the processes
added (celery, celery beat and uwsgi) will be (re)started.

Removed the reread command when restarting supervisor. The update
command rereads all the configurations and adds/removes the processes
as necessary.